+++ ChangeLog.md
@@ -14,6 +14,8 @@ ChangeLog
   * `Sabre\DAV\Locks\Backend\FS`
   * `Sabre\DAV\PartialUpdate\IFile`
   * `Sabre\DAV\URLUtil`
+* Removed: `Sabre\DAV\Client::addTrustedCertificates` and
+  `Sabre\DAV\Client::setVerifyPeer`.
 
 2.1.0-alpha2 (2014-10-23)
 -------------------------